Adjust game speed and movement
Adjust the general game speed and add a slider for contrlling how fast
the game goes. Also scale the speed of the Sym to the game speed and
extracted the function for move towards another object to the GameObject
function. Also makes some other minor adjustments on the decision making
process of Sym.